Sending Swipe Event to trigger sending veggies to Pan
This still needs the animation of the board and maybe an animation so
it doesn’t jump to the pan. Right now it’s really easy to get swipe
events during cutting so often the veggies will jump right to the pan
without the user intentionally swiping. I think when the space key is
pressed it will send an event to the MultiTool to only then listen for
the swiping gesture and that should solve that issue!